# fix(lowest_latency.py): fix latency calc (lower better)